+++ client/client.instrumentation.tests.js
@@ -9,7 +9,7 @@ describe('meteor-coverage', function (done) {
     const downloadPageAndCheckLength = function (i) {
       // meteor-coverage package is a ES6 package
       HTTP.get('/packages/lmieulet_meteor-coverage.js', {}, (error, response) => {
-        if (response.content.length > 14000) {
+        if (response.content.length > 5400) {
           if (i < 10) {
             return downloadPageAndCheckLength(i+1);
           }