+++ server/services/coverage-data.js
@@ -7,6 +7,15 @@ const Report = Npm.require('istanbul-lib-report');
 
 export default CoverageData = {
   getReport: function (coverage) {
+    // QUICKFIX FOR PACKAGES TESTS
+    if (Meteor.isPackageTest) {
+      if (Package["lmieulet:meteor-packages-coverage"] && Package["lmieulet:meteor-packages-coverage"].default && Package["lmieulet:meteor-packages-coverage"].default.CoverageData) {
+        // Ask for lmieulet:meteor-packages-coverage coverage report, as we are in package test
+        return Package["lmieulet:meteor-packages-coverage"].default.CoverageData.getReport(coverage);
+      }
+      throw new Error("lmieulet:meteor-packages-coverage not found. Just add this server dependency in Package.onTest in your package.js")
+    }
+    // Used for meteor apps using babel
     return Coverage.createCoverageMap(coverage);
   },
   getFileReport: function (coverage, filePath) {